requirements.txt
jinja2
langchain==0.3.7
langchain-openai==0.3.3
langchain-community==0.3.7
sentence-transformers
faiss-cpu
PyMuPDF
unstructured
unstructured[pdf]
gunicorn
python-dotenv

flask==2.3.3
flask-cors==4.0.0
openai>=1.58.1,<2.0.0
Pillow==11.0.0
pdf2image==1.16.3
requests==2.31.0
